Base - host all CSS/JS locally to try get MSIE 8 working
MSIE 8 renders everything in the mobile format because respond.js,
which polyfills MSIE 8 to support media queries, does not work
by default with cross-domain CSS. I'm trying to host everything
locally to see if MSIE 8 rendering works.

This is important because a sizable part of the target audience use
MSIE 8.