src/build.py - intelligently handle legacy HTML text
Legacy HTML is also used for descriptions, not just Markdown.
Unfortunately it is often completely invalid, hence we need to use
a very tolerant HTML parser. BeautifulSoup is able to use html5lib
to achieve browser-like parsing.

Additionally we need to fix up image links to be relative in the
correct way.